requirements.txt
asttokens==2.4.1
certifi==2024.8.30
chardet==5.2.0
charset-normalizer==3.4.0
comm==0.2.2
debugpy==1.8.5
decorator==5.1.1
executing==2.0.1
idna==3.10
iniconfig==2.0.0
ipykernel==6.29.5
ipython==8.27.0
jedi==0.19.1
jupyter_client==8.6.2
jupyter_core==5.7.2
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
packaging==24.1
parso==0.8.4
pexpect==4.9.0
pillow==11.0.0
platformdirs==4.2.2
pluggy==1.5.0
prompt_toolkit==3.0.47
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.18.0
pytest==8.3.3
python-dateutil==2.9.0.post0
pyzmq==26.2.0
reportlab==4.2.5
requests==2.32.3
six==1.16.0
stack-data==0.6.3
tornado==6.4.1
traitlets==5.14.3
urllib3==2.2.3
wcwidth==0.2.13
